# docs/OPENAPI_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Healthcare Services Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/verify-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify Otp",
        "operationId": "verify_otp_api_v1_auth_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/firebase/login": {
      "post": {
        "tags": [
          "Firebase Authentication"
        ],
        "summary": "Firebase Login",
        "description": "Exchanges a Firebase ID Token (JWT) from Firebase Phone Auth or Google Sign-In\nfor a platform session access token, auto-provisioning the User profile and\nDPDP consent record on first login.",
        "operationId": "firebase_login_api_v1_auth_firebase_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirebaseLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirebaseLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/firebase/verify": {
      "post": {
        "tags": [
          "Firebase Authentication"
        ],
        "summary": "Verify Token",
        "description": "Verifies a Firebase ID token and inspects its decoded identity claims.",
        "operationId": "verify_token_api_v1_auth_firebase_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirebaseVerifyTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirebaseVerifyTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/firebase/bind-phone": {
      "post": {
        "tags": [
          "Firebase Authentication"
        ],
        "summary": "Bind Phone",
        "description": "Binds or updates a phone number on the currently authenticated user profile.",
        "operationId": "bind_phone_api_v1_auth_firebase_bind_phone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirebaseBindPhoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/firebase/me": {
      "get": {
        "tags": [
          "Firebase Authentication"
        ],
        "summary": "Get Firebase Profile",
        "description": "Returns the authenticated user's Firebase identity profile.",
        "operationId": "get_firebase_profile_api_v1_auth_firebase_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/auth/patient/send-otp": {
      "post": {
        "tags": [
          "Patient Authentication & Profile"
        ],
        "summary": "Send Otp",
        "operationId": "send_otp_api_v1_auth_patient_send_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/patient/verify-otp": {
      "post": {
        "tags": [
          "Patient Authentication & Profile"
        ],
        "summary": "Verify Otp",
        "operationId": "verify_otp_api_v1_auth_patient_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/patient/me": {
      "get": {
        "tags": [
          "Patient Authentication & Profile"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_auth_patient_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Patient Authentication & Profile"
        ],
        "summary": "Update Me",
        "operationId": "update_me_api_v1_auth_patient_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfileRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FirebaseBindPhoneRequest": {
        "properties": {
          "phone_number": {
            "type": "string",
            "title": "Phone Number",
            "description": "E.164 phone number to bind to the user profile"
          }
        },
        "type": "object",
        "required": [
          "phone_number"
        ],
        "title": "FirebaseBindPhoneRequest"
      },
      "FirebaseLoginRequest": {
        "properties": {
          "id_token": {
            "type": "string",
            "title": "Id Token",
            "description": "Firebase ID Token (JWT) acquired from Firebase Client SDK"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Target role (patient, doctor, chemist)",
            "default": "patient"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name",
            "description": "Display name for new patient/user creation"
          },
          "consent_version": {
            "type": "string",
            "title": "Consent Version",
            "description": "DPDP Act patient consent policy version",
            "default": "v1.0"
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "FirebaseLoginRequest"
      },
      "FirebaseLoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "firebase_uid": {
            "type": "string",
            "title": "Firebase Uid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "is_new_user": {
            "type": "boolean",
            "title": "Is New User"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "firebase_uid",
          "user_id",
          "role",
          "is_new_user"
        ],
        "title": "FirebaseLoginResponse"
      },
      "FirebaseVerifyTokenRequest": {
        "properties": {
          "id_token": {
            "type": "string",
            "title": "Id Token",
            "description": "Firebase ID Token to decode and verify"
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "FirebaseVerifyTokenRequest"
      },
      "FirebaseVerifyTokenResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "uid"
        ],
        "title": "FirebaseVerifyTokenResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OTPVerifyRequest": {
        "properties": {
          "phone_number": {
            "type": "string",
            "pattern": "^\\+?[1-9]\\d{9,14}$",
            "title": "Phone Number"
          },
          "otp_code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Otp Code"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "default": "patient"
          },
          "consent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Version",
            "default": "1.0"
          }
        },
        "type": "object",
        "required": [
          "phone_number",
          "otp_code"
        ],
        "title": "OTPVerifyRequest"
      },
      "PatientAuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientProfileRead"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "patient"
        ],
        "title": "PatientAuthResponse"
      },
...
      "PatientProfileRead": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "phone_number"
        ],
        "title": "PatientProfileRead"
      },
      "PatientProfileUpdate": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          }
        },
        "type": "object",
        "title": "PatientProfileUpdate"
      },
...
      "SendOTPRequest": {
        "properties": {
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          }
        },
        "type": "object",
        "required": [
          "phone_number"
        ],
        "title": "SendOTPRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "TokenResponse"
      },
      "UserRead": {
        "properties": {
          "phone_number": {
            "type": "string",
            "pattern": "^\\+?[1-9]\\d{9,14}$",
            "title": "Phone Number",
            "description": "E.164 or Indian mobile format"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "default": "patient"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "phone_number",
          "id",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "UserRead"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "patient",
          "doctor",
          "verification_reviewer",
          "super_admin",
          "chemist"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifyOTPRequest": {
        "properties": {
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "otp_code": {
            "type": "string",
            "title": "Otp Code"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "consent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Version",
            "default": "v1.0"
          }
        },
        "type": "object",
        "required": [
          "phone_number",
          "otp_code"
        ],
        "title": "VerifyOTPRequest"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}